+++ pkg/utils/kubernetes/resources/labels.go
@@ -14,47 +14,36 @@ import (
 // provided object to signal that it's owned by a DataPlane resource
 // and that its lifecycle is managed by this operator.
 func LabelObjectAsDataPlaneManaged(obj metav1.Object) {
-	labels := obj.GetLabels()
-	if labels == nil {
-		labels = make(map[string]string)
-	}
-	labels[consts.GatewayOperatorManagedByLabel] = consts.DataPlaneManagedLabelValue
-	obj.SetLabels(labels)
+	setLabel(obj, consts.GatewayOperatorManagedByLabel, consts.DataPlaneManagedLabelValue)
 }
 
 // LabelObjectAsKongPluginInstallationManaged ensures that labels are set on the
 // provided object to signal that it's owned by a KongPluginInstallation
 // resource and that its lifecycle is managed by this operator.
 func LabelObjectAsKongPluginInstallationManaged(obj metav1.Object) {
-	labels := obj.GetLabels()
-	if labels == nil {
-		labels = make(map[string]string)
-	}
-	labels[consts.GatewayOperatorManagedByLabel] = consts.KongPluginInstallationManagedLabelValue
-	obj.SetLabels(labels)
+	setLabel(obj, consts.GatewayOperatorManagedByLabel, consts.KongPluginInstallationManagedLabelValue)
 }
 
 // LabelObjectAsKonnectExtensionManaged ensures that labels are set on the
 // provided object to signal that it's owned by a KonnectExtension resource
 // and that its lifecycle is managed by this operator.
 func LabelObjectAsKonnectExtensionManaged(obj metav1.Object) {
-	labels := obj.GetLabels()
-	if labels == nil {
-		labels = make(map[string]string)
-	}
-	labels[consts.GatewayOperatorManagedByLabel] = consts.KonnectExtensionManagedByLabelValue
-	obj.SetLabels(labels)
+	setLabel(obj, consts.GatewayOperatorManagedByLabel, consts.KonnectExtensionManagedByLabelValue)
 }
 
 // LabelObjectAsControlPlaneManaged ensures that labels are set on the
 // provided object to signal that it's owned by a ControlPlane resource and that its
 // lifecycle is managed by this operator.
 func LabelObjectAsControlPlaneManaged(obj metav1.Object) {
+	setLabel(obj, consts.GatewayOperatorManagedByLabel, consts.ControlPlaneManagedLabelValue)
+}
+
+func setLabel(obj metav1.Object, key string, value string) { //nolint:unparam
 	labels := obj.GetLabels()
 	if labels == nil {
 		labels = make(map[string]string)
 	}
-	labels[consts.GatewayOperatorManagedByLabel] = consts.ControlPlaneManagedLabelValue
+	labels[key] = value
 	obj.SetLabels(labels)
 }